Add Ch1 checklist implementation
Am I missing something? For whatever reason, the naive implementation,
which is supposed to be n^2, looks like it's just n, while the checklist
implementation is getting wrecked by the naive implementation (it's
consistently about half as fast per execution).

My best guess is GHC is optimizing something in the naive
implementation (but what??), while the two list traversals (the filter
and the takeWhile) are costing the checklist implementation a lot. It
could also be my sparse input (max 10% range coverage) vs. a dense input
maybe.